bit of code refactoring
use multiple context with instead of nesting with

del repo when we're done with it, to recover more memory.

Actually close PE file when done with it.  This isn't so important now
that we're passing in data rather than a file path: it looks like
PE.close does nothing in the data case.