tweak action-download-artifact parameters.
Try to set workflow_conclusion to a falsy value to disable that check
(as opposed to omitting the parameter which defaults it to 'success'

Enable search_artifacts option, so if the latest run doesn't have the
requested artifact it will keep looking for an older one.  I don't know
how the re-running of a single job of a workflow works out, but I could
see it being that only the jobs that actually ran in that run would have
their artifacts associated with that run, and the artifacts from the
other jobs would be on the older run.