track makedepends separately.
Previously, the script would iterate but not recurse makedepends (which
is correct and intended), but if the same package showed up later in the
runtime dependency tree it would be treated as already processed and
still not have its runtime depends recursed.  This feels like a
theoretical concern at this point, but is more likely to cause issues
once a planned change to allow processing multiple starting packages
at once is done.